# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Staff Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Department {
  id: number;
  name: string;
  created_at: string;
  updated_at: string;
}

export interface DepartmentInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name: string;
}

export interface DepartmentUpdate {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name?: string;
}

export interface Role {
  id: number;
  name: string;
}

export interface RoleInput {
  /**
     * @minLength 1
     * @maxLength 50
     */
  name: string;
}

export interface RoleAssignment {
  role_id: number;
}

export interface UserWithRoles {
  id: number;
  full_name: string;
  username: string;
  email: string;
  /** @nullable */
  department_id?: number | null;
  /** @nullable */
  department_name?: string | null;
  created_at: string;
  updated_at: string;
  roles: Role[];
}

export interface UserInput {
  /**
     * @minLength 1
     * @maxLength 150
     */
  full_name: string;
  /**
     * @minLength 1
     * @maxLength 50
     */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
  /** @nullable */
  department_id?: number | null;
  role_ids?: number[];
}

export interface UserUpdate {
  /**
     * @minLength 1
     * @maxLength 150
     */
  full_name?: string;
  /**
     * @minLength 1
     * @maxLength 50
     */
  username?: string;
  email?: string;
  /** @minLength 6 */
  password?: string;
  /** @nullable */
  department_id?: number | null;
  role_ids?: number[];
}

export interface DashboardSummary {
  total_staff: number;
  total_departments: number;
  total_roles: number;
  super_admin_count: number;
}

export interface DepartmentCount {
  department_id: number;
  department_name: string;
  staff_count: number;
}

export interface RoleCount {
  role_id: number;
  role_name: string;
  staff_count: number;
}

export interface Document {
  id: number;
  document_number: string;
  document_date: string;
  subject: string;
  creator_id: number;
  /** @nullable */
  creator_name?: string | null;
  current_status: string;
  file_path: string;
  created_at: string;
  updated_at: string;
}

export interface NextDocumentNumber {
  next_number: string;
}

export interface DocumentInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  document_number: string;
  document_date: string;
  /**
     * @minLength 1
     * @maxLength 255
     */
  subject: string;
  attachment: Blob;
  /** @maxLength 50 */
  current_status?: string;
}

export interface DocumentUpdate {
  /**
     * @minLength 1
     * @maxLength 100
     */
  document_number?: string;
  document_date?: string;
  /**
     * @minLength 1
     * @maxLength 255
     */
  subject?: string;
  /** @maxLength 50 */
  current_status?: string;
}

export interface DocumentLog {
  id: number;
  document_id: number;
  /** @nullable */
  user_id?: number | null;
  /** @nullable */
  user_name?: string | null;
  action: string;
  /** @nullable */
  notes?: string | null;
  timestamp: string;
}

export interface DocumentLogInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  action: string;
  notes?: string;
}

export type ListUsersParams = {
department_id?: number;
role_id?: number;
search?: string;
};

export type ListDocumentsParams = {
search?: string;
status?: string;
};